Add inbound integration test and add JmsReader
Finished the InboundUserAcceptanceTest which sends the EDIFACT test message to MESH mailbox where the adaptor receives inbound transaction.
After that the test compares that the message from inbound queue is the same as the message which has been sent.

Added the JmsReader class which helps us to read the messages.