Remove unused method from IntegrationBaseTest
The getMeshOutboundQueueMessage() method had been added in a
first attempt to update the integration tests to verify the
recep, where I was hoping to read from the outbound queue.
The service picks up the message from the queue and posts it
to the mailbox too quickly though, so this approach was
abandoned and I didn't clear out the then-unused method.