Add segment groups
Segment groups 01, 02, 06, 07, 10, 16, 18, and 20 added as
`InvolvedParty`, `ServiceReportDetails`, `InvestigationSubject`,
`PatientDetails`, `PatientClinicalInfo`, `Specimen`, `LabResult`,
and `ResultReferenceRange`, respectively.

They all extend abstract class `SegmentGroup` which holds a
method for splitting a list of segments into separate lists
according to a segment key used as a separator.

Tests follow the approach that the actual segments within a group
are unimportant at a unit level - only the specific segment keys
that identify a field or segment group.