Add missing FreeText segments
The LabResult segment group (S18) can have up to 99 free text
segments, which may be qualified as either SPC, RIT or CRR.

Only SPC was included, so RIT and CRR needed to be added, as
InvestigationResultFreeText and ComplexReferenceRangeFreeText
respectively.

Since we then had 5 SomethingFreeText classes, all with the same
structure (although some of the existing ones didn't call out
that the text could be in multiple parts), I've added a layer
of abstraction for FreeTextSegments, which they all now extend.

Spotbugs did its usual nonsense with lombok about redundant null
checks, so they have been added to the exclusions list.