Add specimen mappings
Using both the spreadsheet and word document, the mappings
from edifact segments found in SG16 (`SpecimenDetails`) to
HAPIFHIR (`Specimen`) are added.

Still don't know what the URL should be for the fasting
status extension, and we assume that any needed linking to
SG18 (via RFF and SEQ segments) will be done later.