Add MappingUtils to unescape edifact
The raw edifact can include escape sequences
with `?`, e.g. `?'` for the apostrophe, or
`??` for the question mark. The output FHIR
should not retain this format, as it is
edifact-specific. This utility class is
therefore available for mappers to use where
text may need to be unescaped.

SpecimenMapper consumes this for mapping
FreeText segments.