Include reference to patient in each specimen
FHIR Specimens are expected to include a reference to the Patient
from which they were collected, as the "subject" of the resource.
This is meant to be the `fullUrl` of the patient resource, which we
generate from the patient's `id` (a UUID we also generate).

In order to allow the SpecimenMapper to "predict" what the `fullUrl`
will be when defined by the BundleMapper, the responsibility for
determining them has been pulled into a separate component:
ResourceFullUrlGenerator. Now, the edifact mappers and the
BundleMapper can each use the same component to ensure that the
references will match.