NIAD-975: Specimen mapper (#59)
Using both the spreadsheet and word document, the mappings
from edifact segments found in SG16 (`SpecimenDetails`) to
HAPIFHIR (`Specimen`) are added.

The URLs for the "system" fields were taken from the online
examples, but we did not get verification that they are correct
for our service.

MappingUtils is added to unescape edifact. The raw edifact can
include escape sequences with `?`, e.g. `?'` for the apostrophe,
or `??` for the question mark. The output FHIR should not retain
this format, as it is edifact-specific. This utility class is therefore
available for mappers to use where text may need to be
unescaped. SpecimenMapper consumes this for mapping
FreeText segments.

The fasting status extension is part of a different version of FHIR
and can be removed from our codebase.

FHIR Specimens are expected to include a reference to the Patient
from which they were collected, as the "subject" of the resource.
This is meant to be the `fullUrl` of the patient resource, which we
generate from the patient's `id` (a UUID we also generate).

In order to allow the SpecimenMapper to "predict" what the `fullUrl`
will be when defined by the BundleMapper, the responsibility for
determining them has been pulled into a separate component:
ResourceFullUrlGenerator. Now, the edifact mappers and the
BundleMapper can each use the same component to ensure that the
references will match.