Update customizations for date-times
The output fhir json includes ISO8601-formatted date-times, which can be
parsed fine by `OffsetDateTime.parse(String)`, so a customization with that
is added.

There are several places in the integration tests where the same set of customizations are used - so that's been moved into the Base test class.

One integration test uses edifact `DTM+SCO:20100223:102` which does not
include a time, which `OffsetDateTime` cannot parse. But in that case, the
output fhir is string-equal, so the customization is updated to start with
simple string comparison, and only fall back on parsing when necessary.